startup-investor-platform/startup-investor-platform/ml-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3
scipy==1.11.4
python-dotenv==1.0.0
pymongo==4.6.0
httpx==0.25.2
python-multipart==0.0.6
sentence-transformers==2.2.2
joblib==1.3.2
